% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aggregate_data.R
\name{add_column_if_does_not_exist}
\alias{add_column_if_does_not_exist}
\title{Add dummy columns to data.frames, given they do not exist}
\usage{
add_column_if_does_not_exist(df, colnames)
}
\arguments{
\item{df}{data.frame to augment}

\item{colname}{vector of strings; names of potential dummy columns}
}
\description{
Add dummy columns to data.frames, given they do not exist
}
\examples{
df <- data.frame(a = c(1:3, replicate(3,NA)), b = 5:10)
colnames <- c("b", "c", "d")
weatherAndClimateUtils:::add_column_if_does_not_exist(df,colnames)
}
