% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aggregate_data.R
\name{aggregate_data}
\alias{aggregate_data}
\title{Functions for temporally aggregating data or reshaping data into a temporally
aggregated form}
\usage{
aggregate_data(
  data_df,
  type = "synoptic",
  agg_period = "day",
  agg_vars = NULL,
  agg_functions = default_agg_functions()
)
}
\arguments{
\item{data_df;}{data.frame with weather observations from Synoptic Data PBC
or GHCN-Daily.}

\item{type;}{string specifying the data source; so far, "synoptic" and
"ghcnd" are supported}

\item{agg_period;}{string specifying the period to aggregate over; So far, "day",
"month", and "year" are supported.}

\item{agg_functions;}{list of functions for aggregating data}
}
\value{
a data.frame() with observations aggregated to the chosen temporal
resolution with the chosen aggregation functions
}
\description{
Functions for temporally aggregating data or reshaping data into a temporally
aggregated form
}
\examples{
\dontrun{
weatherAndClimatePlots::kska_asos_weather_obs \%>\%
 dplyr::mutate(Date_Time_Local = lubridate::with_tz(Date_Time, tzone = "America/Los_Angeles")) \%>\%
 aggregate_data(type = "synoptic", agg_period = "day") \%>\%
 dplyr::select(tidyselect::all_of(c("date","tmax", "tmin", "precip")))

 weatherAndClimatePlots::san_diego_airport_daily_weather_obs \%>\%
 aggregate_data_ghcnd(agg_period = "wy",
 agg_vars = c("precip", "tmin", "tmax"))
 }
}
