% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aggregate_data.R
\name{aggregate_data_submethods}
\alias{aggregate_data_submethods}
\alias{aggregate_data_subdaily}
\alias{aggregate_data_synoptic}
\title{Submethods for aggregate weather data}
\usage{
aggregate_data_subdaily(
  data_df,
  agg_period = "day",
  agg_vars = NULL,
  agg_functions = default_agg_functions()
)

aggregate_data_synoptic(
  data_df,
  agg_period = "day",
  agg_vars = default_agg_vars_synoptic(),
  agg_functions = default_agg_functions()
)
}
\description{
Submethods for aggregate weather data
}
