% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/weatherAndClimateUtils-package.R
\docType{data}
\name{annual_period_in_days}
\alias{annual_period_in_days}
\title{There are approximately 365.25 days in a year}
\format{
An object of class \code{numeric} of length 1.
}
\usage{
annual_period_in_days
}
\description{
There are approximately 365.25 days in a year
}
\keyword{datasets}
