% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clean_data_ghcnd.R
\name{clean_raw_ghcnd_to_standard}
\alias{clean_raw_ghcnd_to_standard}
\title{GHCND cleaning functions}
\usage{
clean_raw_ghcnd_to_standard(
  raw_ghcnd_df,
  select_SOD_cols = TRUE,
  begin_date = NULL,
  end_date = NULL
)
}
\arguments{
\item{raw_ghcnd_df}{the pure, raw data.frame you get from reading a GHCNd
comma-separated file into R with a function such as read.csv() or
readr::read_csv()}

\item{select_SOD_cols}{\link{logical} Should only the data columns pertaining to
precip, tmin, and tmax which constitute a traditional "summary of the day"
(SOD)?}

\item{begin_date, end_date}{\link[lubridate:posix_utils]{lubridate::POSIXct} (optional) prescribed beginning and end
dates for the record. The default is first and last date contained in
raw_ghcnd_df}
}
\value{
a data.frame with a GHCNd record in a more sensible format; NB, the
default behaviour is such that no data are removed by this function. The
single-character measurement and quality flags are expanded into short
descriptions for easier reading.
}
\description{
Files downloaded from NOAA's Global Historical Climatological Network (GHCN)
daily set are transformed into a more sensible format. Namely,
\enumerate{
\item Units are converted; temperature from tenths of degrees celcius to
celcius and precipitation from tenths of mm to mm
\item The three/four attribute flags for temperature and precipitation
are separated into columns. In the raw GHCN files, the flags of different
types show up as a comma-separated list within a single column. The
three/four flags consist of "measurement", "quality", and "source"
information (with ``time of observation'' occasionally being included as
a fourth element). These are briefly described in the \href{https://www.ncei.noaa.gov/data/global-historical-climatology-network-daily/doc/GHCND_documentation.pdf}{GHCNd Documentation}.
\item column names are made lower-case and precipitation is called "precip"
rather than "prcp"
\item In the case that a day is completely missing from the record (the
row corresponding to a day simply does not exist), a blank/NA record is
added
}
}
\details{
\strong{Quality flags:}
The automated QA flagging procedures are described in \href{https://doi.org/10.1175/2010JAMC2375.1}{Durre et al. 2010}.

One must be cautious about eliminating precipitation values based on outlier
checks or temperature values based on spatial inconsistency checks. All
decisions are application dependent.

\strong{Measurement flags:}
There is one measurement flag which is a complete deal-breaker for
precipitation observations at a daily scale. ``A'' means that a single
value in fact represents a multi-day total.

Flag ``L'' means that a temperature observation is probably offset from its
purported calendar day. This is a dealbreaker for spatial interpolation.

Overall, the QC checks described in Durre et al. 2010 rarely flag valid
observations. However, the level of culling is left to the end user.
}
\examples{
\dontrun{
path <- file.path("","home","alexander","Downloads")
id <- "USC00040983" # Anza Borrego
get_data_ghcnd(dest_dir = path,
               ids = id,
               check_existence = FALSE)
raw_df <- readr::read_csv(file.path(path, paste0(id,".csv")))

clean_raw_ghcnd_to_standard(raw_df)

clean_raw_ghcnd_to_standard(raw_df) \%>\% dplyr::filter(is.na(tmax))

clean_raw_ghcnd_to_standard(raw_df, begin_date = lubridate::ymd("2000-01-01"), end_date = lubridate::ymd("2000-12-31")) \%>\% View()
}
}
