% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aggregate_data.R
\name{aggregate_data_helpers}
\alias{aggregate_data_helpers}
\alias{default_agg_functions}
\alias{default_agg_vars_synoptic}
\alias{synoptic_to_ghcnd_names}
\title{Helper functions and data for aggregate_data()}
\usage{
default_agg_functions()

default_agg_vars_synoptic()

synoptic_to_ghcnd_names(col_name)
}
\arguments{
\item{col_name}{string; column name which will be standardized}
}
\description{
Helper functions and data for aggregate_data()
}
