% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/find_data.R
\name{find_data_synoptic}
\alias{find_data_synoptic}
\title{Check for the existence of files on the system}
\usage{
find_data_synoptic(ids, dir, start, end, format)
}
\arguments{
\item{ids}{vector of strings; synoptic station ids}

\item{dir}{string; the directory to search in}

\item{start, end}{strings; the the start and end times of the record. The
format is YYYYMMDDHHMM in Universal Coordinated Time [UTC]}

\item{format}{string; file format: csv or json}
}
\value{
a vector of strings with filepaths of the files on the system which
match the given description
}
\description{
Check for the existence of files on the system
}
\details{
If given format is "csv", then each station ID is iterated through.

If the given format is "json", only json files which contain \emph{all} station
IDs will come up in the search.
}
