% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clean_data_ghcnd.R
\name{GHCNd_wide_format_to_standard_summary_of_day}
\alias{GHCNd_wide_format_to_standard_summary_of_day}
\title{Convert NOAA's ".dly" fixed width file format to their '3-flag' style csv
format}
\usage{
GHCNd_wide_format_to_standard_summary_of_day(wide_fwf_in, SOD_csv_out = NULL)
}
\arguments{
\item{wide_fwf_in}{\link{character} the path to a NOAA ".dly" fixed width file}

\item{SOD_csv_out}{\link{character} the path to write a NOAA 3-flag style csv}
}
\value{
a data.frame reminiscient of the format or nothing/NULL, if a SOD_csv_out
is provided.

If a data.frame() is returned, this can be fed to clean_raw_ghcnd_to_standard()
}
\description{
Convert NOAA's ".dly" fixed width file format to their '3-flag' style csv
format
}
\examples{
\dontrun{
path <- file.path(
"","home","alexander","data", "observations",
"ghcn-daily", "subsets", "socal-temperature-qc",
"qc1out")

id <- "USC00040983"

wide_fwf_in <- file.path(path, paste0(id, ".dly"))

#SOD_csv_out

NOAA_3flag_df <- GHCNd_wide_format_to_standard_summary_of_day(wide_fwf_in = wide_fwf_in)

clean_raw_ghcnd_to_standard(NOAA_3flag_df) \%>\%
 dplyr::filter(date > lubridate::ymd("20260501"), date < lubridate::ymd("20260520"))
}

}
