% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/seasonal_cycle.R
\name{harmonic_quantile_regression}
\alias{harmonic_quantile_regression}
\title{Harmonic Quantile Regression method for calculating seasonal cycles}
\usage{
harmonic_quantile_regression(df, quantile = 0.5, n_harmonics = 2)
}
\arguments{
\item{df}{a data.frame() of daily weather observations, with the first column
containing DATE or POSIXct objects and the remaining columns containing the
weather obs (e.g. maximum temperature, precipitation).

Remove all rows with NAs before calculating. I will fix bad behaviour on my
end}

\item{quantile}{quantiles for which we want a seasonal cycle; common ones
include 0.05, 0.5 (the median), and 0.95}

\item{n_harmonics}{how many cosine waves we wish to superimpose; two might be
too coarse and six might be too noisy}
}
\value{
a data.frame() with the original variable and its quantiles
}
\description{
This function calculates quantiles conditioned on the time of year.
}
\details{
For example, if you wanted to know the median of the daily maximum
temperature on 30 June, you could take the median of all observed maximum
temperatures on 30 June, or a narrow window of days about 30 June. This
approach does not make the best use of available data and the results are
needlessly messy. If, however, we accept that annual cycles are, by
definition, periodic functions, we can instead regress on the time of year.
We use the function quantreg::rq() for this purpose.

This function should not be called directly. Instead, use seasonal_cycle()
from this package.
}
