% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/seasonal_cycle.R
\name{harmonic_quantile_regression_core}
\alias{harmonic_quantile_regression_core}
\title{Core calculation function for harmonic_quantile_regression()}
\usage{
harmonic_quantile_regression_core(df, quantile, variable_name)
}
\arguments{
\item{df}{a data.frame() of daily weather observations, with the first column
containing DATE or POSIXct objects and the remaining columns containing the
weather obs (e.g. maximum temperature, precipitation)}

\item{quantile}{quantiles for which we want a seasonal cycle; common ones
include 0.05, 0.5 (the median), and 0.95}

\item{variable_name}{name of variable for which we calcualte the seasonal cycle}
}
\value{
a data.frame() with values of a variable and its conditional quantiles
}
\description{
Core calculation function for harmonic_quantile_regression()
}
