% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clean_data_ghcnd.R
\name{standard_summary_of_day_to_GHCNd_wide_format}
\alias{standard_summary_of_day_to_GHCNd_wide_format}
\title{Convert "human standard" tidy summary of the day to the ".dly" format of NOAA}
\usage{
standard_summary_of_day_to_GHCNd_wide_format(
  SOD_in_df,
  ghcn_id,
  out_dir,
  elements_v = c("tmin", "tmax", "precip")
)
}
\arguments{
\item{SOD_in_df}{\link{data.frame} data.frame with the summary of the day}

\item{ghcn_id}{\link{character} the GHCN ID for the station}

\item{out_dir}{\link{character} the directory to write the file}

\item{elements_v}{\link{character} vector the variables to include. For now, this
is set to c("tmin", "tmax", "precip")}
}
\value{
a data.frame reminiscient of the format or nothing/NULL, if a out_dir
is provided
}
\description{
Some programs such as NOAA's QA/QC routine expect data to be written in a
rather wide format. This format is described in a NOAA readme file (ftp://ftp.ncdc.noaa.gov/pub/data/ghcn/daily/readme.txt), which is
somewhat reproduced here.
}
\details{
\subsection{General format is as follows:}{
}

\subsection{Variable   Columns   Type}{

ID            1-11   Character
YEAR         12-15   Integer
MONTH        16-17   Integer
ELEMENT      18-21   Character
VALUE1       22-26   Integer
MFLAG1       27-27   Character
QFLAG1       28-28   Character
SFLAG1       29-29   Character
VALUE2       30-34   Integer
MFLAG2       35-35   Character
QFLAG2       36-36   Character
SFLAG2       37-37   Character
.           .          .
.           .          .
.           .          .
VALUE31    262-266   Integer
MFLAG31    267-267   Character
QFLAG31    268-268   Character
SFLAG31    269-269   Character

Now, it is assumed that the units of SOD_in_df are celcius for temperature and
mm for precipitation.

The output is written in tenths of mm and tenths of degrees C (as integers)

It is also assumed that at least some month has in the range has a 31st. This
will be true for any dataset more than a couple of months long.#'
}
}
