% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_data_synoptic.R
\name{synoptic}
\alias{synoptic}
\alias{get_data_synoptic}
\alias{build_synoptic_api_call}
\alias{get_api_key_synoptic}
\alias{set_api_key_synoptic}
\alias{get_clean_and_save_data_synoptic}
\title{Download station observations from Synoptic Data PBC}
\usage{
get_data_synoptic(
  dest_dir,
  api_key = NULL,
  ids,
  start = NULL,
  end = NULL,
  recent = NULL,
  format = "csv"
)

build_synoptic_api_call(
  api_key = NULL,
  ids,
  start = NULL,
  end = NULL,
  recent = NULL,
  format = "csv"
)

get_api_key_synoptic()

set_api_key_synoptic(given_api_key, force = FALSE)

get_clean_and_save_data_synoptic(
  dest_dir_raw_data,
  dest_file_processed_data = NULL,
  api_key = NULL,
  ids,
  start = NULL,
  end = NULL,
  recent = NULL,
  format = "json"
)
}
\arguments{
\item{dest_dir}{string, the directory to download to}

\item{api_key}{string, API Token obtained from Synoptic Data PBC. A free
account is required. See https://docs.synopticdata.com/account/public-api-tokens
for more information}

\item{ids}{vector of strings, the IDs of the stations from which you want to
download observations}

\item{start, end}{strings, the boundaries of the time interval over which you
wish to fetch observations formatted as "yyyymmddhhmm". The timezone is
universal coordinated time (UTC), so be sure to consider offset from your
local time}

\item{recent}{numeric, integer specifically; Instead of providing an interval,
you can specify the "recent", the number of minutes you wish to count back from
the present moments. Set recent=120 to download the most recent two hours of
obs.}

\item{format}{string; file format for downloaded data. The valid options are
"json" and "csv".}

\item{given_api_key}{string of the API token associated with your
synoptic account}

\item{force}{logical; if TRUE, the user consents to writing the key to a file
which persists between R sessions; if FALSE, the user is asked for consent to
write.}

\item{dest_dir_raw_data}{same as dest_dir in get_data_synoptic}

\item{dest_file_processed_data}{string; the desired name of the standardized
synoptic tibble, including its file path}
}
\value{
NULL; the function is called for its side-effect
}
\description{
With a basic, free account, anybody can download up to a year of station
observations from Synoptic.
}
\examples{
get_data_synoptic(dest_dir = "~/Downloads", ids = c("KSAN", "G3667"),
start = "202308080000", end = "202308090000")

get_clean_and_save_data_synoptic(dest_dir_raw_data = file.path("~", "Downloads"),
dest_file_processed_data = file.path("~","data","weather_today.Rda"),
ids = c("KSAN", "G3667", "G4747"),
recent = 120)
}
