% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aggregate_data.R
\name{tag_wy}
\alias{tag_wy}
\title{Tag water year}
\usage{
tag_wy(year, month, wy_first_month = 10)
}
\arguments{
\item{year}{numeric; year}

\item{month}{numeric; month}

\item{wy_first_month}{numeric; the month which defines the beginning of the
water year. In the western United States, this is 10 (1 October).}
}
\description{
Tag water year
}
