% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/join_data.R
\name{merge_data}
\alias{merge_data}
\title{Functions for consolidating observations from different data sources}
\usage{
merge_data(synoptic_df = NULL, synoptic_daily_df = NULL, ghcnd_df = NULL)
}
\arguments{
\item{synoptic_df}{data.frame; observations from Synoptic Data PBC, as
outputted by weatherAndClimateUitls::load_data_synoptic()}

\item{synoptic_daily_df}{data.frame; observations from synoptic, summarized
to daily resolution by weatherAndClimateUtils::aggregate_data()}

\item{ghcnd_df}{data.frame; observations from GHCNd, as outputted by
weatherAndClimatePlots::clean_raw_ghcnd_to_standard()}
}
\value{
a data.frame() with GHCNd observations augmented with observations
from Synoptic. Any observations which exist in GHCNd take precidence, because
their maxima and minima are more trustworthy (since they are not as dependent)
on the temporal resolution of the full hi-res data as the synoptic obs, which
are sometimes hourly.
}
\description{
Functions for consolidating observations from different data sources
}
