% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/seasonal_cycle.R
\name{seasonal_cycle}
\alias{seasonal_cycle}
\title{Calculate Seasonal Cycles from daily weather data}
\usage{
seasonal_cycle(df, .cols, method = "harmonic_quantile_regression", ...)
}
\arguments{
\item{df}{a data.frame() of daily weather observations, with the first column
containing DATE or POSIXct objects and the remaining columns containing the
weather obs (e.g. maximum temperature, precipitation)}

\item{.cols}{a character vector listing variables for which we want to calculate
the seasonal cycle}

\item{method}{string of the particular method used to calculate seasonal cycles;
currently, only "harmonic_quantile_regression" is supported}

\item{...}{extra arguments passed to the chosen method}
}
\value{
a data.frame() with the date, the chosen variables, and their annual
cycles
}
\description{
Calculate Seasonal Cycles from daily weather data
}
